backend/requirements.txt
flask
flask-cors
Pillow
python-dotenv
uuid
werkzeug